src/requirements.txt
asgiref==3.8.1
Django==4.2.20
djangorestframework==3.16.0
numpy==2.0.2
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
